requirements.txt
absl-py==1.4.0
aiosignal==1.3.1
alabaster==0.7.13
antlr4-python3-runtime
appdirs
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
Babel==2.12.1
beautifulsoup4
biopython
bokeh
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer
click
cloudpickle
cmake==3.26.0
colorama==0.4.6
configparser==5.3.0
contourpy
cryptography==41.0.1
csvsort==1.6.1
cycler
Cython
cytoolz==0.12.1
dask
Distance==0.1.3
distlib==0.3.6
distributed
docutils==0.19
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.9.1
flatbuffers==23.5.26
fonttools
frozenlist==1.3.3
fsspec
future
gast==0.4.0
gdown==4.6.4
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
GraKeL==0.1.9
grpcio==1.51.3
h5py==3.8.0
HeapDict==1.0.1
hydra-core
icecream==2.1.3
idna
imagesize==1.4.1
importlib-metadata==6.0.0
importlib-resources
iniconfig==2.0.0
jax==0.4.11
Jinja2
joblib==1.2.0
jsonschema==4.17.3
kaleido==0.2.1
keras==2.12.0
kiwisolver
libclang==16.0.0
lit==15.0.7
locket
logging-exceptions==0.1.8
lz4
Markdown==3.4.3
MarkupSafe
matplotlib
matplotlib-venn==0.11.9
ml-dtypes==0.1.0
modin==0.19.0
mpmath==1.3.0
msgpack
munch==2.5.0
mxfold2
networkx
nltk==3.8.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf
OpenMM==8.0.0
opt-einsum==3.3.0
packaging
pandas==1.5.3
partd
pdbfixer
Pillow
platformdirs
plotly==5.13.1
pluggy==1.1.0
pooch
progressbar2==4.2.0
protobuf==4.22.1
psutil
pyaml==21.10.1
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycparser==2.21
Pygments==2.14.0
pynisher==1.0.2
pyOpenSSL==23.2.0
pyparsing
pyrsistent==0.19.3
PySocks
pytest==7.3.2
python-dateutil
python-utils==3.5.2
pytz
PyYAML
ray
regex==2022.10.31
requests
requests-oauthlib==1.3.1
rna-tools==3.12.1
rsa==4.9
scikit-learn==1.2.2
scipy
six
snowballstemmer==2.2.0
sortedcontainers
soupsieve==2.4.1
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-autoprogram==0.1.8
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sympy==1.11.1
tblib
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
tomli==2.0.1
toolz
torch==1.13.1
torchvision==0.14.1
tornado
tqdm==4.65.0
triton==2.0.0
typing_extensions
tzdata
unicodedata2
UpSetPlot==0.8.0
urllib3==1.26.16
versioneer==0.28
virtualenv==20.21.0
Werkzeug==2.3.4
wget==3.2
wrapt==1.14.1
xyzservices
zict
zipp
